refactor: Unify Radix Wallet Links
This commit goes through the hard work of consolidating any references
to the main `radix` sdk into a single `useWallet`.  There have been
problems with inconsistent ledger behavior while processing transactions
of all forms while swapping between active accounts and physical ledger
devices.

Originally this larger scope of work had been built on the premise that
the radix sdk must be `activated` with the referenced active account
just before a required action was performed.  It seems as though the
SDK's state does not transfer around between references, callback and
promises; nor should we expect it to.

This commit unifes those interactions by:

- removing all references to `activateAccount` from outside of the
  `useWallet` API.
- refactor the behavior of `useTransactions` into the main `useWallet`
  API.  This allows us to only reference the main SDK, and no longer
  depend on intermediate references that seem to eventually fall out of
  sync.
- refactor `useHistory` to move decryption of messages back to the
  `useWallet` sdk.  `useHistory` still serves as a holder of decrypted
  messages.
- Since `activateAccount` is now internal to `useWallet`, call it before
  `transfer`, `stake` or `unstake` actions.  `activateAccount` now acts
  only on the main `radix` instance and we can no longer attempt to
  inject new Wallets into the SDK (I'd guess this was the source of the
  bugs)
- Guard against double clicks on long-running account actions by adding
  a `isActivating` boolean that discards activation attempts while an
  activation is ongoing
