+++ src/composables/useHistory.ts
@@ -17,7 +17,24 @@ const isDecrypting: Ref<boolean> = ref(false)
 let transactionSub: Subscription | null
 
 const activeAddress: Ref<AccountAddressT | null> = ref(null)
-export default function useHistory (radix: ReturnType<typeof Radix.create>, address: AccountAddressT) {
+
+interface useHistoryInterface {
+  canGoBack: ComputedRef<boolean>;
+  canGoNext: Ref<boolean>;
+  decryptedMessages: Ref<{id: string, message: string}[]>;
+  isDecrypting: Ref<boolean>;
+  loadingHistory: Ref<boolean>;
+  transactions: Ref<SimpleExecutedTransaction[]>;
+  pushMsg: (tx: ExecutedTransaction, msg: string) => void;
+  fetchTransactions: (cursor?: string) => Promise<void>;
+  leavingHistory: () => void;
+  nextPage: () => void;
+  previousPage: () => void;
+  resetHistory: () => void;
+  updateActiveAccount: (addr: AccountAddressT) => void;
+}
+
+export default function useHistory (radix: ReturnType<typeof Radix.create>): useHistoryInterface {
   const fetchTransactions = async (cursor?: string) => {
     if (!activeAddress.value) return
     const params = { size: PAGE_SIZE, address: activeAddress.value, cursor }
@@ -51,13 +68,8 @@ export default function useHistory (radix: ReturnType<typeof Radix.create>, addr
     activeAddress.value = addr
   }
 
-  const decryptMessage = (client: ReturnType<typeof Radix.create>, tx: ExecutedTransaction) => {
-    isDecrypting.value = true
-    firstValueFrom(client.decryptTransaction(tx))
-      .then((val) => {
-        decryptedMessages.value.push({ id: tx.txID.toString(), message: val })
-      })
-      .finally(() => { isDecrypting.value = false })
+  const pushMsg = (tx: ExecutedTransaction, msg: string) => {
+    decryptedMessages.value.push({ id: tx.txID.toString(), message: msg })
   }
 
   const previousPage = () => {
@@ -84,7 +96,7 @@ export default function useHistory (radix: ReturnType<typeof Radix.create>, addr
     decryptedMessages,
     loadingHistory,
     transactions,
-    decryptMessage,
+    pushMsg,
     fetchTransactions,
     leavingHistory,
     nextPage,
